3d_figs/requirements.txt
pipreqs # to get useful packages
ipython==8.12.0
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
Pillow==9.5.0
plotly==5.14.1
scipy==1.10.1
